WEEK 3/May 21/requirements.txt
autogen-core==0.5.7
autogen-agentchat==0.5.7
google-generativeai==0.8.5
pandas==2.2.3
matplotlib==3.10.3
